fix: disable deleting my hashtags (#889)
## Summary

I ran into some sync issues, not caused by this but it does cascade to
deleting my favorite node and I think we should disable it.

## How to test


## Checklist

- [ ] I followed the [style
guide](https://www.notion.so/Style-guide-08041425021b49b28a0734f4e8105c33).
- [ ] I tested my contribution.
- [ ] I added new environment variables to
[Vercel](https://vercel.com/ideaflowco/mew/settings/environment-variables).

+++ src/db/graphNodes.ts
@@ -91,6 +91,9 @@ export const deleteNode = async (tx: MewDbTransaction, node: SerializedNode) => 
   if (node.id.startsWith(USER_MY_HASHTAGS_NODE_ID_PREFIX)) {
     throw new SyncError('Cannot delete user\'s "My Hashtags" node', { actionName: "deleteNode", data: { node } });
   }
+  if (node.id.startsWith(USER_MY_FAVORITES_NODE_ID_PREFIX)) {
+    throw new SyncError('Cannot delete user\'s "My Favorites" node', { actionName: "deleteNode", data: { node } });
+  }
 
   // Delete all relationLists entries that reference this node
   await tx.delete(relationListsTable).where(eq(relationListsTable.nodeId, node.id));
